% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alpha}
\alias{alpha}
\title{This function computes the misclassification error \code{alpha}.}
\usage{
alpha(Delta, tauSq, Y, X, num_Pred)
}
\arguments{
\item{Delta}{Empirical Bayes  effect size estimates from \code{EBayes}}

\item{tauSq}{Empirical Bayes variance estimates from \code{EBayes}}

\item{Y}{list of response variables as explained in \code{EBayes}}

\item{X}{A merge of the orginal predictors values used in computing the z-values.
The  dataset with response variable as first entry of the list \code{Y} stays ontop
and the second dataset is that with response variable as second entry of the list \code{Y} and so on.}

\item{num_Pred}{is the number of predictor from which the prediction error is computed.}
}
\description{
\code{alpha} returns  alpha_pos and alpha_neg and their values after correlation
correction
}
