% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Y.R
\docType{data}
\name{Y}
\alias{Y}
\title{Binary response variable of four studies with correponding z-values in the Z dataset}
\format{An object of class \code{list} of length 4.}
\usage{
Y
}
\description{
A list containing the response variables of four studies. All response variables are
binary and coded as 0 and 1
}
\keyword{datasets}
