% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z.R
\docType{data}
\name{Z}
\alias{Z}
\title{z-values of 54659 genes from 4 studies}
\format{An object of class \code{matrix} with 54659 rows and 4 columns.}
\usage{
Z
}
\description{
A dataset containing the z-values of four studies. Each column contains the z-values of
each study.
}
\keyword{datasets}
